+++ index.html
@@ -3,7 +3,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
   <meta name="viewport" content="width=device-width,initial-scale=1">
   
-<script src="./scripts.js" type="text/javascript"></script>
+<script src="./scripts/scripts.js" type="module"></script>
 <link href="./styles.css" rel="stylesheet" type="text/css" />
 </head>
