+++ topics/enhanced.js
@@ -1,0 +1,192 @@
+export const enhancedSticks = [
+    'Step 1',
+    'Step 2',
+    'Step 3',
+    'Step 4',
+    'Step 5',
+    'Step 6',
+    'Step 7',
+    'Step 8',
+    'Step 9',
+    'Step 10',
+    'Step 11',
+    'Step 12',
+    'Fear',
+    'Warped Lives',
+    'The Promises', 
+    'Guidance',
+    'Service Work',
+    'Easier, Softer Way',
+    'Isolating',
+    'Easy Does It',
+    'Faith',
+    'Drop The Rock',
+    'Wreckage of Past',
+    'Financial Amends',
+    'God',
+    'Honesty',
+    'Humility',
+    'The New Comer',
+    'Desperation',
+    'Helping Others',
+    'Bargaining',
+    'There is a Solution',
+    'Serenity',
+    'Going to Any Length',
+    'Willingness',
+    'Lack of Control',
+    'Your First 30 Days',
+    'Confrontation',
+    'Peace',
+    'Sponsorship',
+    'Self Pity',
+    'Hopelessness',
+    'Fellowship',
+    'Courage',
+    'Living Sober',
+    'Spiritual Growth',
+    'Attraction not Promotion',
+    'Higher Power',
+    'Meditation',
+    'Healing Families',
+    'Working Program at Home',
+    'Judgemental',
+    'Gods Will',
+    'Love',
+    'Sanity',
+    'Bitterness',
+    'Demoralization',
+    'Trusted Servants',
+    'Truth',
+    'Recovery',
+    'Half Measures',
+    'Hope',
+    'Resentments',
+    'Spiritual Progress',
+    'Living Amends',
+    'Patience',
+    'Anger',
+    'One Day at a Time',
+    'Economic Insecurity',
+    'Rigorous Honesty',
+    'Jealousy',
+    'Family Disease',
+    'Self-Seeking',
+    'Gratitude',
+    'Joyous',
+    'Meeting Crisis',
+    'Sunlight of the Spirit',
+    'First Things First',
+    'Change',
+    'Trudging Buddies',
+    'Controversy',
+    'Self Loathing',
+    'Spiritual Experience',
+    'Common Solution',
+    'Ego',
+    'Failure',
+    'Retaliation',
+    'Unity',
+    'Spiritual Adviser',
+    'Third Step Prayer',
+    'Trust',
+    'Cunning, Baffling, Powerful',
+    // Group-added topics:
+    'Finding the solution',
+    'What is your favorite step (and why)?',
+    'Which of the promises have come true for you?',
+    'What was your miracle?',
+    "Life on life's terms",
+    "How do you handle your triggers?",
+    "Strategies between meetings",
+    "Dealing with big emotions",
+    'Right-Sized',
+    'The Traditions',
+    'Emotional Sobriety',
+    'Rule 62',
+    'Expectations',
+    'Can the order in which the Steps are worked be changed?',
+    // bonus topics (source: https://www.aa.org/assets/en_us/smf-56_en.pdf):
+    'Acceptance',
+    'Attitude of Gratitude',
+    'Complacency',
+    'Contempt Prior to Investigation',
+    'Dependence',
+    'Forgiveness',
+    'Freedom Through Sobriety',
+    'Inadequacy',
+    'Inventory',
+    'Letting go of Anger',
+    'Making Amends',
+    'Open-mindedness',
+    'Participation and Action',
+    'Patience and Tolerance',
+    'Plan the Action not the Result',
+    'Practice these principles in all our affairs',
+    'Principles before personalities',
+    'Projection - living in the wreckage of the future',
+    'Staying away from the first drink',
+    'Surrender',
+    'Three Legacies - Recovery, Unity and Service',
+    '"Twelfth stepping"',
+    'Ways of carrying the AA message',
+    'What is sobriety?',
+    'Working with others',
+    // Bonus topics from emailed source: "101 Meeting Starters A Guide to Better Twelve Step Discussions - Mel B."
+    'Willpower',
+    'Are Alcoholics Perfectionists?',
+    'Are We Passing It On?',
+    'Attracting Trouble',
+    'Be Careful What You Pray For',
+    'Being Responsible',
+    'Changing Things We Can',
+    'Coming to Grips with Fear',
+    'Contending with self-will',
+    'Controlling the Imagination',
+    'Coping with Depression',
+    'Coping with Social Pressure',
+    'Dealing with Disagreeable People',
+    'Dealing with Rejection',
+    'Dealing with the Past',
+    'Do Material Things Matter?',
+    'Do We Deserve Success?',
+    'Fearing Change',
+    'Erasing the Old Tapes',
+    'Finding True Independence',
+    'Getting Beyond People Pleasing',
+    'Giving Away to Keep',
+    'Growth through Prayer and Action',
+    'Happy coincidences',
+    'How to Deal with Impatience',
+    'How to Find Happy Sobriety',
+    'Hungry, Angry, Lonely, Tired',
+    'Keep it Simple',
+    'Letting go of guilt',
+    'Life after cloud nine',
+    'Needing the program vs wanting it',
+    'Old resentments flaring up',
+    'The need for self honesty',
+    'Wanting instant gratification',
+    'We die to live',
+    'What about my old friends?',
+    'What blocks acceptance?',
+    'What is being spiritually fit?',
+    'What is Prayer & Meditation?',
+    'What is sincerity?',
+    'When have we made a decision?',
+    'Right action leads to right thinking',
+    'When pride gets in my way',
+    'When the worst things become the best',
+    'Who is an alcoholic?',
+    'Whom can we fix?',
+    'Why attitude matters',
+    'Why did we drink?',
+    'Why do I keep coming back?',
+    'Why have a primary purpose?',
+    'Why help is needed',
+    'Why is one drink so bad?',
+    'Why should we make amends?',
+    'Why willingness is the key',
+    'Winning the boredom battle',
+    'Courage to change the things we can',
+];
