+++ topics/original.js
@@ -1,0 +1,94 @@
+export const originalSticks = [
+    'Step 1',
+    'Step 2',
+    'Step 3',
+    'Step 4',
+    'Step 5',
+    'Step 6',
+    'Step 7',
+    'Step 8',
+    'Step 9',
+    'Step 10',
+    'Step 11',
+    'Step 12',
+    'Fear',
+    'Warped Lives',
+    'The Promises', 
+    'Guidance',
+    'Service Work',
+    'Easier, Softer Way',
+    'Isolating',
+    'Easy Does It',
+    'Faith',
+    'Drop The Rock',
+    'Wreckage of Past',
+    'Financial Amends',
+    'God',
+    'Honesty',
+    'Humility',
+    'The New Comer',
+    'Desperation',
+    'Helping Others',
+    'Bargaining',
+    'There is a Solution',
+    'Serenity',
+    'Going to Any Length',
+    'Willingness',
+    'Lack of Control',
+    'Your First 30 Days',
+    'Confrontation',
+    'Peace',
+    'Sponsorship',
+    'Self Pity',
+    'Hopelessness',
+    'Fellowship',
+    'Courage',
+    'Living Sober',
+    'Spiritual Growth',
+    'Attraction not Promotion',
+    'Higher Power',
+    'Meditation',
+    'Healing Families',
+    'Working Program at Home',
+    'Judgemental',
+    'Gods Will',
+    'Love',
+    'Sanity',
+    'Bitterness',
+    'Demoralization',
+    'Trusted Servants',
+    'Truth',
+    'Recovery',
+    'Half Measures',
+    'Hope',
+    'Resentments',
+    'Spiritual Progress',
+    'Living Amends',
+    'Patience',
+    'Anger',
+    'One Day at a Time',
+    'Economic Insecurity',
+    'Rigorous Honesty',
+    'Jealousy',
+    'Family Disease',
+    'Self-Seeking',
+    'Gratitude',
+    'Joyous',
+    'Meeting Crisis',
+    'Sunlight of the Spirit',
+    'First Things First',
+    'Change',
+    'Trudging Buddies',
+    'Controversy',
+    'Self Loathing',
+    'Spiritual Experience',
+    'Common Solution',
+    'Ego',
+    'Failure',
+    'Retaliation',
+    'Unity',
+    'Spiritual Adviser',
+    'Third Step Prayer',
+    'Trust',
+    'Cunning, Baffling, Powerful',
+];
