+++ topics/weekly.js
@@ -1,0 +1,36 @@
+export const weeklyApplication = [
+	"Share what you're looking forward to about a random step (or your favorite part of one you've already finished)",
+	"Is every area of your life manageable? How did you, or didn't you, practice step 1 today?",
+	"What are you currently working on (or resisting) giving to your HP?",
+	"In what ways were you, or could you be, further 'restored to sanity' this week?",
+	"In what ways did you (or did you not) turn your will & life over to your HP today?",
+	"Where did you see your HP show up today/this week?",
+	"Are there any recent resentments or fears in which you struggle to see your part? How do you handle that?",
+	"This week, how did you (or could you) practice taking a 'fearless moral inventory'?",
+	"In what ways has your sponsor or another member of the fellowship helped you this week?",
+	"Are there any 'defects' of character you're not yet ready, or took longer to become ready, to have HP remove?",
+	"How did you (or will you) practice humbly asking HP to remove your shortcomings this week?",
+	"Which of the 9th step promises showed up most for you this week?",
+	"How often do you 'continue to take personal inventory'? What is your process?",
+	"What did your practice of prayer & meditation look like this week? (We are not saints - how could it improve?)",
+	"This week, how has HP given you the 'knowledge of HP's will for us and the power to carry it out'?",
+	"How did you work your program in your daily life this week?",
+	"Fears are big & small - how have you found courage despite fear recently?",
+	"It requires no willingness to do something we WANT to do... in what area(s) have you practiced the most willingness this week?",
+	"In what ways did you, or could you, avoid isolation & encourage true connection this week?",
+	"Where were you able to practice 'easy does it' this week? Where might you practice it this coming week?",
+	"How does your biggest worry today compare to your biggest worry from last week? month? 5 years ago?",
+	"What's one character *asset* of yours that showed up this week, and how?",
+	"What big or small moment brought you the most serenity this week?",
+	"If someone else struggled with your biggest worry, what could you say that might reframe it, and/or inspire hope?",
+	"In what ways did you try to control things this week? How did you practice letting go of control?",
+	"What can you appreciate most... about the person or situation you liked the least this week?",
+	"In what ways did 'esteemable acts' (big or small) help your self esteem this week?",
+	"How did you practice 'first things first' this week?",
+	"What does it mean to 'meet crisis with serenity'? How have you, or might you, do this with a recent 'crisis' (big or small)?",
+	"How did you, or could you, practice contrary action this week?", 
+	"How have you been recently reminded you are not the Ultimate Authority?",
+	// "How did you, or could you, practice being 'fully self supporting' (regardless of finances) this week?",
+	"Where did you, or could you, practice 'Rule 62' this week (not take yourself too damn seriously)? ",
+	"Where did you practice acceptance this week? Where was acceptance more challenging?",
+];
